+++ Changelog.md
@@ -8,6 +8,7 @@ Lilu Changelog
 - Added centralised entitlement hooking API
 - Added lilu_os_qsort export (the supported interface is Apple-private)
 - Added `liludelay=1000` boot argument to insert a 1s delay after each print
+- Added new symbol routing API with simplified interface
 - Fixed a kernel panic in userspace patching code on 10.14b1
 
 #### v1.2.3